change editor file hierarchy
