+++ assets/src/components/editor/Editor.tsx
@@ -8,8 +8,8 @@ import {
 } from 'data/content/model';
 import { editorFor, markFor } from './editors';
 import { ToolbarItem, CommandContext } from './commands/interfaces';
-import { onKeyDown as listOnKeyDown } from './editors/Lists';
-import { commandDesc as linkCmd } from 'components/editor/commands/CmdLink';
+import { onKeyDown as listOnKeyDown } from './editors/lists/ListsOverride';
+import { commandDesc as linkCmd } from 'components/editor/commands/LinkCmd';
 import { getNearestBlock } from './utils';
 import { toggleMark } from './commands/commands';
 import { installNormalizer } from './normalizer';
