+++ assets/src/components/editor/editors/code/CodeSettings.tsx
@@ -1,19 +1,13 @@
 import React, { useState, useRef, useEffect } from 'react';
-import { ReactEditor } from 'slate-react';
-import { Transforms } from 'slate';
-import { updateModel, getEditMode } from './utils';
 import * as ContentModel from 'data/content/model';
-import { EditorProps, CommandContext } from './interfaces';
+import { CommandContext } from 'components/editor/editors/interfaces';
 import guid from 'utils/guid';
-import * as Settings from './Settings';
+import * as Settings from 'components/editor/editors/settings/Settings';
 
 const languages = Object
   .keys(ContentModel.CodeLanguages)
   .filter(k => typeof ContentModel.CodeLanguages[k as any] === 'number')
   .sort();
-
-export interface CodeProps extends EditorProps<ContentModel.Code> {
-}
 
 type CodeSettingsProps = {
   model: ContentModel.Code,
@@ -23,7 +17,7 @@ type CodeSettingsProps = {
   editMode: boolean,
 };
 
-const CodeSettings = (props: CodeSettingsProps) => {
+export const CodeSettings = (props: CodeSettingsProps) => {
 
   // Which selection is active, URL or in course page
   const [model, setModel] = useState(props.model);
@@ -101,57 +95,4 @@ const CodeSettings = (props: CodeSettingsProps) => {
       </div>
     </div>
   );
-};
-
-export const CodeEditor = (props: CodeProps) => {
-
-  const { editor } = props;
-  const { model } = props;
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-
-  const editMode = getEditMode(editor);
-
-  const onEdit = (updated: ContentModel.Code) => {
-    updateModel<ContentModel.Code>(editor, model, updated);
-    setIsPopoverOpen(false);
-  };
-
-  const onRemove = () => {
-    ($('#remove-button') as any).tooltip('hide');
-
-    const path = ReactEditor.findPath(editor, model);
-    Transforms.removeNodes(editor, { at: path });
-
-    setIsPopoverOpen(false);
-  };
-
-  const contentFn = () => <CodeSettings
-    model={model}
-    editMode={editMode}
-    commandContext={props.commandContext}
-    onRemove={onRemove}
-    onEdit={onEdit} />;
-
-  return (
-    <div {...props.attributes} className="code-editor">
-      <div className="code-editor-content">
-        <pre style={{ fontFamily: 'Menlo, Monaco, Courier New, monospace' }} >
-          <code className={`language-${model.language}`}>{props.children}</code>
-        </pre>
-      </div>
-
-      <div contentEditable={false} style={{ userSelect: 'none' }}>
-        <Settings.ToolPopupButton
-          contentFn={contentFn}
-          setIsPopoverOpen={setIsPopoverOpen}
-          isPopoverOpen={isPopoverOpen}
-          label="Code" />
-        <Settings.Caption caption={model.caption} />
-      </div>
-    </div>
-  );
-};
-
-export const CodeBlockLine = (props: any) => {
-  return <div {...props.attributes}>{props.children}</div>;
 };
