+++ assets/src/components/editor/editors/link/ExistingLink.tsx
@@ -1,35 +1,8 @@
 import React, { useState, useEffect, useRef } from 'react';
-import Popover from 'react-tiny-popover';
-import * as ContentModel from 'data/content/model';
-import { ReactEditor } from 'slate-react';
-import { Transforms } from 'slate';
-import { EditorProps, CommandContext } from './interfaces';
-import { updateModel } from './utils';
-import { Action, onEnterApply } from './Settings';
 import * as Persistence from 'data/persistence/resource';
-
-const internalLinkPrefix = '/course/link';
-
-const isInternalLink = (href: string) => href.startsWith(internalLinkPrefix);
-
-const isValidHref = (href: string) => href.startsWith('https://')
-  || href.startsWith('http://')
-  || href.startsWith('mailto://')
-  || href.startsWith('ftp://');
-
-const addProtocol = (href: string) => isValidHref(href)
-  ? href
-  : 'http://' + href;
-
-export const normalizeHref = (href: string) => addProtocol(href.trim());
-
-// Takes a delivery oriented internal link and translates it to
-// a link that will resolve at authoring time. This allows
-// authors to use the 'Open Link' function and visit the linked course
-// page.
-const translateDeliveryToAuthoring = (href: string, projectSlug: string) => {
-  return `/project/${projectSlug}/resource/` + href.substr(href.lastIndexOf('/') + 1);
-};
+import { CommandContext } from 'components/editor/commands/interfaces';
+import { onEnterApply, Action } from 'components/editor/editors/settings/Settings';
+import { internalLinkPrefix, normalizeHref } from './utils';
 
 type ExistingLinkProps = {
   href: string,
@@ -53,9 +26,7 @@ interface Uninitialized {
 
 type LinkablePages = Uninitialized | Waiting | Persistence.PagesReceived;
 
-
-
-const ExistingLink = (props: ExistingLinkProps) => {
+export const ExistingLink = (props: ExistingLinkProps) => {
 
   // Which selection is active, URL or in course page
   const [isURL, setIsURL] = useState(true);
@@ -245,78 +216,5 @@ const ExistingLink = (props: ExistingLinkProps) => {
         {input}
       </div>
     </div>
-  );
-};
-
-
-export interface LinkProps extends EditorProps<ContentModel.Hyperlink> {
-}
-
-export const LinkEditor = (props: LinkProps) => {
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const { attributes, children, editor, model } = props;
-
-  const onEdit = (href: string) => {
-
-    if (href !== '' && href !== model.href) {
-      updateModel<ContentModel.Hyperlink>(editor, model, { href });
-    }
-    setIsPopoverOpen(false);
-  };
-
-  const onVisit = (href: string) => {
-
-    if (isInternalLink(href)) {
-      window.open(translateDeliveryToAuthoring(
-        href, props.commandContext.projectSlug), '_blank');
-    } else {
-      window.open(href, '_blank');
-    }
-
-  };
-
-  const onCopy = (href: string) => {
-
-    if (isInternalLink(href)) {
-      navigator.clipboard.writeText(
-        window.location.protocol + '//' + window.location.host + '/' +
-        translateDeliveryToAuthoring(href, props.commandContext.projectSlug));
-    } else {
-      navigator.clipboard.writeText(href);
-    }
-
-  };
-
-  const onRemove = () => {
-    ($('#remove-button') as any).tooltip('hide');
-
-    const path = ReactEditor.findPath(editor, model);
-    Transforms.unwrapNodes(editor, { at: path });
-
-    setIsPopoverOpen(false);
-  };
-
-  return (
-    <a id={props.model.id} href="#"
-      className="inline-link" {...attributes} onClick={() => setIsPopoverOpen(true)}>
-      <Popover
-        onClickOutside={() => {
-          setIsPopoverOpen(false);
-        }}
-        isOpen={isPopoverOpen}
-        padding={25}
-        position={['bottom', 'top', 'left', 'right']}
-        content={() => <ExistingLink
-          href={model.href}
-          commandContext={props.commandContext}
-          onVisit={onVisit}
-          onCopy={onCopy}
-          onRemove={onRemove}
-          onChange={() => { }}
-          inModal={false}
-          onEdit={onEdit} />}>
-        {ref => <span ref={ref}>{children}</span>}
-      </Popover>
-    </a>
   );
 };
