+++ assets/src/components/editor/editors/settings/Settings.scss
