+++ assets/src/components/editor/toolbars/fixed/fixedToolbarItems.ts
@@ -1,7 +1,7 @@
-import { commandDesc as imgCmdDesc } from 'components/editor/commands/CmdImage';
-import { commandDesc as ytCmdDesc } from 'components/editor/commands/CmdYouTube';
-import { commandDesc as audioCmdDesc } from 'components/editor/commands/CmdAudio';
-import { commandDesc as tableCommandDesc } from 'components/editor/commands/CmdTable';
+import { commandDesc as imgCmdDesc } from 'components/editor/commands/ImageCmd';
+import { commandDesc as ytCmdDesc } from 'components/editor/commands/YoutubeCmd';
+import { commandDesc as audioCmdDesc } from 'components/editor/commands/AudioCmd';
+import { commandDesc as tableCommandDesc } from 'components/editor/commands/table/TableCmd';
 import { ResourceType } from 'data/content/resource';
 import { ToolbarItem } from 'components/editor/commands/interfaces';
